Author: cyli
Reviewer: therve, exarkun
Fixes #3097

Change FilePath.remove so that symlinks are only deleted by it, not the contents
of the directories they point to, and so that when a symlink which points to a
directory is encountered, an exception is not raised.


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@23062 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/python/filepath.py
@@ -499,7 +499,13 @@ class FilePath(_PathHelper):
         utime(self.path, None)
 
     def remove(self):
-        if self.isdir():
+        """
+        Removes the file or directory that is represented by self.  If
+        C{self.path} is a directory, recursively remove all its children
+        before removing the directory.  If it's a file or link, just delete
+        it.
+        """
+        if self.isdir() and not self.islink():
             for child in self.children():
                 child.remove()
             os.rmdir(self.path)
